Fix format specifier warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36503 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/supersnapshot.c
@@ -173,7 +173,7 @@ static void supersnapshot_v5_io1_store(uint16_t addr, uint8_t value)
 
 static int supersnapshot_v5_dump(void)
 {
-    mon_out("Register: $%02x (%s)\n", currreg, (ss_rom_disabled) ? "disabled" : "enabled");
+    mon_out("Register: $%02x (%s)\n", (unsigned int)currreg, (ss_rom_disabled) ? "disabled" : "enabled");
     mon_out(" EXROM: %d GAME: %d (%s)\n", ((romconfig >> 1) & 1), (romconfig & 1) ^ 1, cart_config_string((uint8_t)(romconfig & 3)));
     mon_out(" ROM %s, Bank: %d\n", (ss_rom_disabled) ? "disabled" : "enabled", currbank);
     mon_out(" RAM %s, Bank: %d\n", (export_ram) ? "enabled" : "disabled", ram_bank);
